requirements.txt
apscheduler>=3.6.3
flask-login>=0.6.3
flask-wtf>=1.2.2
psycopg2-binary>=2.9.10
twilio>=9.4.6
apscheduler>=3.6.3
flask
flask-sqlalchemy
pytelegrambotapi
wtforms
twilio
psutil
flask-login
flask-wtf
twilio
flask-login
flask-wtf
sqlalchemy
werkzeug
python-dotenv
python-telegram-bot==13.7
telegram
flask
flask-login
flask-sqlalchemy
flask-wtf
psycopg2-binary
pytelegrambotapi
python-telegram-bot==13.7
flask-login
flask-wtf
flask
flask-login
flask-sqlalchemy
flask-wtf
psycopg2-binary
pytelegrambotapi
python-telegram-bot==13.7
trafilatura
flask-login
flask-sqlalchemy
werkzeug
pillow
flask-login
flask-sqlalchemy
psycopg2-binary
python-dotenv
werkzeug
bbcode
Pillow
bbcode
pillow
flask-migrate
flask-sqlalchemy
psycopg2-binary
markdown2
mistune
bbcode
flask-login
flask-migrate
flask-sqlalchemy
flask-login
flask-wtf
